make config loadable, add server run and stop fns
- configuration is loadable from a configuration file
- service can be started with `Registrator.RegServer.main()` method
